add lib module to `src/tool/compiletest`
